+++ src/extensions/fixtures.ts
@@ -24,24 +24,24 @@ export const flatten = (ob: any) => {
 
 export type KeysToValues<
   T extends object,
-  Ks extends (keyof T)[],
+  Ks extends readonly (keyof T)[],
 > = Ks extends [
   infer F1 extends keyof T,
   ...infer Rest extends (keyof T)[],
 ]
   ? Rest extends []
-    ? [T[F1]]
+    ? T[F1]
     : [T[F1], ...KeysToValues<T, Rest>]
   : [];
 
-export function mapArray<K extends string>(
+export function mapArray<const K extends string>(
   key: K,
 ): <const T extends Record<K, any>>(value: T) => T[K];
 
 export function mapArray<
-  K1 extends string,
-  K2 extends string,
-  Ks extends string[],
+  const K1 extends string,
+  const K2 extends string,
+  const Ks extends string[],
 >(
   key1: K1,
   key2: K2,
@@ -50,11 +50,11 @@ export function mapArray<
   value: T,
 ) => [T[K1], T[K2], ...KeysToValues<T, Ks>];
 
-export function mapArray<const T extends object>(...keys: (keyof T)[]) {
-  return (value: T) => {
+export function mapArray<T extends readonly string[]>(...keys: T) {
+  return (value: Record<T[number], any>) => {
     const len = keys.length;
-    if (len === 1) return value[keys[0]];
-    return keys.map(key => value[key]);
+    if (len === 1) return (value as any)[keys[0]];
+    return keys.map(key => (value as any)[key]);
   };
 }